+++ tagmanager/ctags/read.h
@@ -106,7 +106,7 @@ extern void fileUngetc (int c);
 extern const unsigned char *fileReadLine (void);
 extern char *readLine (vString *const vLine, MIO *const mio);
 extern char *readSourceLine (vString *const vLine, MIOPos location, long *const pSeekValue);
-extern boolean bufferOpen (unsigned char *buffer, int buffer_size,
+extern boolean bufferOpen (unsigned char *buffer, size_t buffer_size,
 			   const char *const fileName, const langType language );
 #define bufferClose fileClose
